Separate secret/vault provider into its own package
Move the secret/vault-provider classes (controllers, service, mapper,
entity, repository, converters, encryption utils) and their tests into
com.otilm.common.credential.provider.secret.*, leaving the credential
provider and connector-wide classes (Info/Health controllers) at the
root. Pure package + import move; Spring component scanning is unchanged
(everything stays under the base package) and no behavior changes.
