--- src/main/java/com/otilm/common/credential/provider/api/SecretTypeConverter.java
@@ -1,14 +1,0 @@
-package com.otilm.common.credential.provider.api;
-
-import com.otilm.api.model.connector.secrets.SecretType;
-
-import java.beans.PropertyEditorSupport;
-
-public class SecretTypeConverter extends PropertyEditorSupport {
-
-    @Override
-    public void setAsText(String text) throws IllegalArgumentException {
-        setValue(SecretType.findByCode(text));
-    }
-
-}
