+++ src/main/java/com/otilm/common/credential/provider/secret/api/SecretTypeConverter.java
@@ -1,0 +1,14 @@
+package com.otilm.common.credential.provider.secret.api;
+
+import com.otilm.api.model.connector.secrets.SecretType;
+
+import java.beans.PropertyEditorSupport;
+
+public class SecretTypeConverter extends PropertyEditorSupport {
+
+    @Override
+    public void setAsText(String text) throws IllegalArgumentException {
+        setValue(SecretType.findByCode(text));
+    }
+
+}
