--- src/main/java/com/otilm/common/credential/provider/util/SecretEncodingVersion.java
@@ -1,14 +1,0 @@
-package com.otilm.common.credential.provider.util;
-
-import lombok.Getter;
-
-@Getter
-public enum SecretEncodingVersion {
-    V1("v1");
-
-    private final String version;
-
-    SecretEncodingVersion(String version) {
-        this.version = version;
-    }
-}
