+++ src/test/java/com/otilm/common/credential/provider/secret/dao/converter/SecretTypeConverterTest.java
@@ -1,0 +1,22 @@
+package com.otilm.common.credential.provider.secret.dao.converter;
+
+import com.otilm.api.model.connector.secrets.SecretType;
+import com.otilm.common.credential.provider.secret.api.SecretTypeConverter;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+import java.util.List;
+
+class SecretTypeConverterTest {
+
+    private final SecretTypeConverter converter = new SecretTypeConverter();
+
+    @Test
+    void testConvert() {
+        List<String> secretCodes = Arrays.stream(SecretType.values()).map(SecretType::getCode).toList();
+        for (String code : secretCodes) {
+            Assertions.assertDoesNotThrow(() -> converter.setAsText(code));
+        }
+    }
+}
